# docs/specs/api-contract.yaml
...
openapi: 3.0.3
info:
  title: ORBIT Backend API
  version: 1.0.0
  description: >
    Contract for the ORBIT FastAPI backend. Dashboard and Hermes are both
    clients of this API. The API exposes state and actions only — never rules;
    all conditional/triggered logic lives in n8n. Every state-changing endpoint
    writes an audit_log entry attributed to the calling actor (user | hermes | n8n).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /rooms:
    get:
      tags: [rooms]
      summary: List rooms (with device counts)
      responses:
        '200':
          description: Rooms
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Room' }
    post:
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /rooms/{roomId}:
    parameters:
      - name: roomId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [rooms]
      summary: Get a room
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [rooms]
      summary: Update a room
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [rooms]
      summary: Delete a room (devices are unassigned, not deleted)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------- scenes ----
  /scenes:
    get:
      tags: [scenes]
      summary: List scenes
      responses:
        '200':
          description: Scenes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Scene' }
    post:
      tags: [scenes]
      summary: Create a scene (a static, ordered list of capability writes — no conditions)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SceneCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scene' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /scenes/{sceneId}:
    parameters:
      - name: sceneId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [scenes]
      summary: Get a scene with its actions
      responses:
        '200':
          description: Scene
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scene' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [scenes]
      summary: Update a scene and/or replace its actions
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SceneUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scene' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [scenes]
      summary: Delete a scene
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /scenes/{sceneId}/execute:
    post:
      tags: [scenes]
      summary: Execute a scene's actions in order
      description: >
        Callable by dashboard users, Hermes, and n8n workflows. This endpoint
        performs actions only — deciding *when* to call it is n8n's job.
      parameters:
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: Execution started; per-action results reported
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneExecutionResult' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------- audit ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Users authenticate via /auth/login. Hermes and n8n use long-lived
        service principals (JWTs with actor_type claims 'hermes' / 'n8n')
        issued from backend config — the audit log reads the actor from
        this claim.
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: device_offline }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Room:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        icon: { type: string, nullable: true }
        sort_order: { type: integer }
        device_count: { type: integer, readOnly: true }
    RoomCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string, description: Generated from name if omitted }
        icon: { type: string }
        sort_order: { type: integer }
    RoomUpdate:
      type: object
      properties:
        name: { type: string }
        icon: { type: string }
        sort_order: { type: integer }
    SceneAction:
      type: object
      required: [device_id, capability, payload]
      properties:
        device_id: { type: string, format: uuid }
        capability: { type: string }
        payload:
          type: object
          example: { value: false }
        sort_order: { type: integer }
    Scene:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        icon: { type: string, nullable: true }
        description: { type: string, nullable: true }
        actions:
          type: array
          items: { $ref: '#/components/schemas/SceneAction' }
    SceneCreate:
      type: object
      required: [name, actions]
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }
        actions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SceneAction' }
    SceneUpdate:
      type: object
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }
        actions:
          description: Full replacement of the action list when provided
          type: array
          items: { $ref: '#/components/schemas/SceneAction' }
    SceneExecutionResult:
      type: object
      properties:
        scene_id: { type: string, format: uuid }
        audit_id: { type: string, format: uuid }
        results:
          type: array
          items:
            type: object
            properties:
              device_id: { type: string, format: uuid }
              capability: { type: string }
              status: { type: string, enum: [published, skipped_offline, error] }
...